big changes with Teasy element to better support non-existing element case
