+++ src/main/java/com/wiley/autotest/selenium/context/AbstractPageElement.java
@@ -9,6 +9,7 @@ import com.wiley.autotest.selenium.ParamsProvider;
 import com.wiley.autotest.selenium.Report;
 import com.wiley.autotest.selenium.elements.CheckBox;
 import com.wiley.autotest.selenium.elements.upgrade.TeasyElement;
+import com.wiley.autotest.selenium.elements.upgrade.VisibleTeasyElement;
 import com.wiley.autotest.selenium.elements.upgrade.Window;
 import com.wiley.autotest.selenium.extensions.internal.DefaultElementFactory;
 import com.wiley.autotest.spring.Settings;
@@ -548,9 +549,9 @@ public abstract class AbstractPageElement<P extends AbstractPageElement> extends
     @Deprecated
     //Will be removed from framework. Do not use this method. Create something similar in your project instead.
     //TODO VE - delete this by Sept 2017
-    public List<String> getTextFromWebElementList(final List<TeasyElement> webElementList) {
+    public List<String> getTextFromWebElementList(final List<VisibleTeasyElement> webElementList) {
         final List<String> resultList = new ArrayList<String>();
-        for (TeasyElement eachElement : webElementList) {
+        for (VisibleTeasyElement eachElement : webElementList) {
             scrollIntoView(eachElement);
             resultList.add(eachElement.getText().trim());
         }
@@ -697,7 +698,7 @@ public abstract class AbstractPageElement<P extends AbstractPageElement> extends
 
     @Deprecated
     /**
-     * use {@link TeasyElement#getParent()}
+     * use {@link VisibleTeasyElement#getParent()}
      */
     public WebElement getParentElement(final WebElement webElement) {
         return getParentElement(webElement, 1);
@@ -705,7 +706,7 @@ public abstract class AbstractPageElement<P extends AbstractPageElement> extends
 
     @Deprecated
     /**
-     * use {@link TeasyElement#getParent()}
+     * use {@link VisibleTeasyElement#getParent()}
      */
     public WebElement getParentElement(final WebElement webElement, int level) {
         StringBuilder builder = new StringBuilder(".");
