+++ src/main/java/com/wiley/autotest/selenium/elements/AbstractWebContainer.java
@@ -4,8 +4,8 @@ import com.wiley.autotest.WebDriverAwareElementFinder;
 import com.wiley.autotest.selenium.SeleniumHolder;
 import com.wiley.autotest.selenium.context.TeasyElementProvider;
 import com.wiley.autotest.selenium.context.ErrorSender;
-import com.wiley.autotest.selenium.elements.upgrade.CustomWebElement;
 import com.wiley.autotest.selenium.elements.upgrade.TeasyElement;
+import com.wiley.autotest.selenium.elements.upgrade.VisibleTeasyElement;
 import org.openqa.selenium.Alert;
 import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
@@ -65,8 +65,8 @@ public abstract class AbstractWebContainer extends TeasyElementProvider implemen
 
     private static void castToCustomWebElement(Object[] args) {
         for (int i = 0; i < args.length; i++) {
-            if (args[i] instanceof CustomWebElement) {
-                args[i] = ((CustomWebElement) args[i]).getWrappedWebElement();
+            if (args[i] instanceof TeasyElement) {
+                args[i] = ((TeasyElement) args[i]).getWrappedWebElement();
             }
         }
     }
