+++ src/main/java/com/wiley/autotest/selenium/elements/upgrade/v3/conditions/element/ElementAttributeNotContain.java
@@ -1,4 +1,4 @@
-package com.wiley.autotest.selenium.elements.upgrade.v3.expectedconditions;
+package com.wiley.autotest.selenium.elements.upgrade.v3.conditions.element;
 
 import com.wiley.autotest.selenium.elements.upgrade.TeasyElement;
 import org.openqa.selenium.WebDriver;
@@ -7,14 +7,15 @@ import org.openqa.selenium.support.ui.ExpectedCondition;
 import javax.annotation.Nullable;
 
 /**
- * Created by vefimov on 23/08/2017.
+ * Created by vefimov on 24/08/2017.
  */
-public class ElementAttributeValue implements ExpectedCondition<Boolean> {
+public class ElementAttributeNotContain implements ExpectedCondition<Boolean> {
+
     private TeasyElement element;
     private String attributeName;
     private String value;
 
-    public ElementAttributeValue(TeasyElement element, String attributeName, String value) {
+    public ElementAttributeNotContain(TeasyElement element, String attributeName, String value) {
         this.element = element;
         this.attributeName = attributeName;
         this.value = value;
@@ -23,11 +24,12 @@ public class ElementAttributeValue implements ExpectedCondition<Boolean> {
     @Nullable
     @Override
     public Boolean apply(@Nullable WebDriver driver) {
-        return element.getAttribute(attributeName).equals(value);
+        String attribute = element.getAttribute(attributeName);
+        return attribute != null && !attribute.contains(value);
     }
 
     @Override
     public String toString() {
-        return String.format("Element '%s' attribute '%s' value is not '%s'", element.toString(), attributeName, value);
+        return String.format("Element '%s' attribute '%s' contains value '%s'!", element.toString(), attributeName, value);
     }
 }
