+++ src/main/java/com/wiley/autotest/selenium/elements/upgrade/v3/conditions/element/ElementAttributeValue.java
@@ -1,4 +1,4 @@
-package com.wiley.autotest.selenium.elements.upgrade.v3.expectedconditions;
+package com.wiley.autotest.selenium.elements.upgrade.v3.conditions.element;
 
 import com.wiley.autotest.selenium.elements.upgrade.TeasyElement;
 import org.openqa.selenium.WebDriver;
@@ -7,15 +7,14 @@ import org.openqa.selenium.support.ui.ExpectedCondition;
 import javax.annotation.Nullable;
 
 /**
- * Created by vefimov on 24/08/2017.
+ * Created by vefimov on 23/08/2017.
  */
-public class ElementAttributeNotContain implements ExpectedCondition<Boolean> {
-
+public class ElementAttributeValue implements ExpectedCondition<Boolean> {
     private TeasyElement element;
     private String attributeName;
     private String value;
 
-    public ElementAttributeNotContain(TeasyElement element, String attributeName, String value) {
+    public ElementAttributeValue(TeasyElement element, String attributeName, String value) {
         this.element = element;
         this.attributeName = attributeName;
         this.value = value;
@@ -24,12 +23,11 @@ public class ElementAttributeNotContain implements ExpectedCondition<Boolean> {
     @Nullable
     @Override
     public Boolean apply(@Nullable WebDriver driver) {
-        String attribute = element.getAttribute(attributeName);
-        return attribute != null && !attribute.contains(value);
+        return element.getAttribute(attributeName).equals(value);
     }
 
     @Override
     public String toString() {
-        return String.format("Element '%s' attribute '%s' contains value '%s'!", element.toString(), attributeName, value);
+        return String.format("Element '%s' attribute '%s' value is not '%s'", element.toString(), attributeName, value);
     }
 }
