+++ src/main/java/com/wiley/autotest/selenium/elements/upgrade/v3/conditions/element/ElementHasAttribute.java
@@ -1,4 +1,4 @@
-package com.wiley.autotest.selenium.elements.upgrade.v3.expectedconditions;
+package com.wiley.autotest.selenium.elements.upgrade.v3.conditions.element;
 
 import com.wiley.autotest.selenium.elements.upgrade.TeasyElement;
 import org.openqa.selenium.WebDriver;
@@ -7,14 +7,13 @@ import org.openqa.selenium.support.ui.ExpectedCondition;
 import javax.annotation.Nullable;
 
 /**
- * Created by vefimov on 24/08/2017.
+ * Created by vefimov on 23/08/2017.
  */
-public class ElementNotHaveAttribute implements ExpectedCondition<Boolean> {
-
+public class ElementHasAttribute implements ExpectedCondition<Boolean> {
     private TeasyElement element;
     private String attributeName;
 
-    public ElementNotHaveAttribute(TeasyElement element, String attributeName) {
+    public ElementHasAttribute(TeasyElement element, String attributeName) {
         this.element = element;
         this.attributeName = attributeName;
     }
@@ -22,12 +21,11 @@ public class ElementNotHaveAttribute implements ExpectedCondition<Boolean> {
     @Nullable
     @Override
     public Boolean apply(@Nullable WebDriver driver) {
-        return element.getAttribute(attributeName) == null;
+        return element.getAttribute(attributeName) != null;
     }
 
     @Override
     public String toString() {
-        return String.format("Element '%s' has attribute '%s'", element.toString(), attributeName);
+        return String.format("Element '%s' does not have attribute '%s'", element.toString(), attributeName);
     }
-
 }
