+++ src/main/java/com/wiley/autotest/selenium/extensions/internal/AbstractEnabledElement.java
@@ -2,6 +2,7 @@ package com.wiley.autotest.selenium.extensions.internal;
 
 import com.wiley.autotest.selenium.elements.EnabledElement;
 import com.wiley.autotest.selenium.elements.upgrade.TeasyElement;
+import com.wiley.autotest.selenium.elements.upgrade.VisibleTeasyElement;
 import org.openqa.selenium.By;
 
 abstract class AbstractEnabledElement extends AbstractElement implements EnabledElement {
