+++ src/main/java/com/wiley/autotest/services/Configuration.java
@@ -1,6 +1,7 @@
 package com.wiley.autotest.services;
 
 import com.wiley.autotest.selenium.elements.upgrade.TeasyElement;
+import com.wiley.autotest.selenium.elements.upgrade.VisibleTeasyElement;
 import org.openqa.selenium.remote.DesiredCapabilities;
 import org.springframework.stereotype.Component;
