+++ src/main/java/com/wiley/autotest/spring/SeleniumTestExecutionListener.java
@@ -252,13 +252,15 @@ public class SeleniumTestExecutionListener extends AbstractTestExecutionListener
                 SeleniumHolder.setIOSDriver(iosDriver);
             }
 
+            //TODO VE, NT - consider reworking with new model in mind (Visible/Dom/Null)
+
             String classFromSettings = settings.getProperty("our.webelement.class");
             if (classFromSettings != null && !classFromSettings.isEmpty()) {
                 SeleniumHolder.setOurWebElementClass(classFromSettings);
             } else if (configuration.getClassOfElement() != null) {
                 SeleniumHolder.setOurWebElementClass(configuration.getClassOfElement().getName());
             } else {
-                SeleniumHolder.setOurWebElementClass("com.wiley.autotest.selenium.elements.upgrade.TeasyElement");
+                SeleniumHolder.setOurWebElementClass("com.wiley.autotest.selenium.elements.upgrade.VisibleTeasyElement");
             }
         }
     }
@@ -678,7 +680,7 @@ public class SeleniumTestExecutionListener extends AbstractTestExecutionListener
         capabilities.setCapability(CapabilityType.ACCEPT_SSL_CERTS, true);
         setAlertBehaviorCapabilities(capabilities);
         setProxy(capabilities);
-        //Found that setting this capability could increase IE tests speed. Should be checked.
+        //Found that setting this capability could increase IE tests speed. VisibleShould be checked.
         capabilities.setCapability(InternetExplorerDriver.REQUIRE_WINDOW_FOCUS, true);
         return capabilities;
     }
