+++ src/test/java/com/wiley/autotest/framework/tests/elements/NonexistentElementThrowsException.java
@@ -2,6 +2,7 @@ package com.wiley.autotest.framework.tests.elements;
 
 import com.wiley.autotest.framework.config.BaseTest;
 import com.wiley.autotest.framework.pages.TestElementPage;
+import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.TimeoutException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.testng.annotations.Test;
@@ -14,7 +15,7 @@ public class NonexistentElementThrowsException extends BaseTest {
     @Autowired
     private TestElementPage testElementPage;
 
-    @Test(expectedExceptions = AssertionError.class)
+    @Test(expectedExceptions = NoSuchElementException.class)
     public void test() {
         openPage("mainTestElement.html");
         testElementPage.checkElementNotFound();
